pyproject.toml
[project]
name = "test-fastapi-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi-mcp>=0.2.0",
    "langchain>=0.3.23",
    "langchain-mcp-adapters>=0.0.8",
    "langchain-openai>=0.3.12",
    "langgraph>=0.3.30",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.1",
]
